fix: styling change to make text opaque and padding changes
